Transformer and Shortcode -- Added config
Added support for passing a) a configuration hash to the transformer b) a configuration object to shortcodes. As a first implementation, the whitelisted shortcodes can be configured through a config hash. Default behavior for whitelisting is, that if the whitelist is not overridden, all available shortcodes in registered bundles are whitelisted. Updated tests.